chore: move Attendance Model to Domain
